requirements.txt
seleniumwire
Flask==1.1.2
pbkdf2==1.3
pycryptodome==3.9.8
pyodbc==4.0.30
requests==2.23.0
selenium==3.141.0
selenium-wire==2.1.2
SQLAlchemy==1.3.18
webdriver-manager==3.2.1
webdriverdownloader==1.1.0.3

